refactor(home): relocate OS4 home hooks into feature packages
Move the OS4-only home entry points out of the version-first `os4/`
folders so the version tag lives in the class name instead, matching
the repo convention:

- app/os4/home/{HomePad,HomePhone}.java -> app/Home/os4/
  (package ...libhook.app.Home.os4)
- rules/home/os4/HomeSettingsOS4.kt -> rules/home/other/
- rules/home/os4/NativeHomeHooks.kt -> rules/home/other/NativeHomeHooksOS4.kt

NativeHomeHooks is renamed to NativeHomeHooksOS4 because its two
`external fun`s are resolved by their JNI symbol names, which encode the
fully-qualified class name (there is no RegisterNatives table to keep
the two in sync). The matching exports in
app/src/main/cpp/hyperceiler_home.cpp are updated in the same commit.

Also drop the dead stub rules/home/dock/DockAutoHideOS4.kt (0
references) and harmonize the touched files' license headers with the
repo's dominant full-AGPL header.
