+++ library/libhook/src/main/java/com/sevtinge/hyperceiler/libhook/app/Home/os4/HomePhone.java
@@ -1,13 +1,22 @@
 /*
  * This file is part of HyperCeiler.
- *
+
  * HyperCeiler is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License.
- *
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License.
+
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
  * Copyright (C) 2023-2026 HyperCeiler Contributions
  */
-package com.sevtinge.hyperceiler.libhook.app.os4.home;
+package com.sevtinge.hyperceiler.libhook.app.Home.os4;
 
 import com.hchen.database.HookBase;
 import com.sevtinge.hyperceiler.common.utils.PrefsBridge;
@@ -15,14 +24,14 @@ import com.sevtinge.hyperceiler.libhook.base.BaseLoad;
 import com.sevtinge.hyperceiler.libhook.rules.home.HomePortraitReverse;
 import com.sevtinge.hyperceiler.libhook.rules.home.gesture.HomeGestureOS4;
 import com.sevtinge.hyperceiler.libhook.rules.home.gesture.ShakeDevice;
-import com.sevtinge.hyperceiler.libhook.rules.home.os4.HomeSettingsOS4;
-import com.sevtinge.hyperceiler.libhook.rules.home.os4.NativeHomeHooks;
+import com.sevtinge.hyperceiler.libhook.rules.home.other.HomeSettingsOS4;
+import com.sevtinge.hyperceiler.libhook.rules.home.other.NativeHomeHooksOS4;
 
-@HookBase(targetPackage = "com.miui.home", deviceType = 1, minOSVersion = 4.0F)
-public class HomePad extends BaseLoad {
+@HookBase(targetPackage = "com.miui.home", deviceType = 2, minOSVersion = 4.0F)
+public class HomePhone extends BaseLoad {
     @Override
     public void onPackageLoaded() {
-        NativeHomeHooks.INSTANCE.init();
+        NativeHomeHooksOS4.INSTANCE.init();
         initHook(HomeSettingsOS4.INSTANCE, true);
         boolean gesturesEnabled = PrefsBridge.getBoolean("home_gesture_enable");
         boolean hasTouchGesture = PrefsBridge.getInt("home_gesture_double_tap_action", 0) > 0
