--- library/libhook/src/main/java/com/sevtinge/hyperceiler/libhook/rules/home/dock/DockAutoHideOS4.kt
@@ -1,1 +1,0 @@
-// Removed: OS4 Flutter launcher has no Java folder/editPanel methods
