+++ library/libhook/src/main/java/com/sevtinge/hyperceiler/libhook/rules/home/other/HomeSettingsOS4.kt
@@ -1,8 +1,22 @@
 /*
- * This file is part of HyperCeiler.
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-package com.sevtinge.hyperceiler.libhook.rules.home.os4
+  * This file is part of HyperCeiler.
+
+  * HyperCeiler is free software: you can redistribute it and/or modify
+  * it under the terms of the GNU Affero General Public License as
+  * published by the Free Software Foundation, either version 3 of the
+  * License.
+
+  * This program is distributed in the hope that it will be useful,
+  * but WITHOUT ANY WARRANTY; without even the implied warranty of
+  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  * GNU Affero General Public License for more details.
+
+  * You should have received a copy of the GNU Affero General Public License
+  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+  * Copyright (C) 2023-2026 HyperCeiler Contributions
+*/
+package com.sevtinge.hyperceiler.libhook.rules.home.other
 
 import android.app.Activity
 import android.content.res.Configuration
