+++ library/libhook/src/main/java/com/sevtinge/hyperceiler/libhook/rules/home/other/NativeHomeHooksOS4.kt
@@ -1,8 +1,22 @@
 /*
- * This file is part of HyperCeiler.
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-package com.sevtinge.hyperceiler.libhook.rules.home.os4
+  * This file is part of HyperCeiler.
+
+  * HyperCeiler is free software: you can redistribute it and/or modify
+  * it under the terms of the GNU Affero General Public License as
+  * published by the Free Software Foundation, either version 3 of the
+  * License.
+
+  * This program is distributed in the hope that it will be useful,
+  * but WITHOUT ANY WARRANTY; without even the implied warranty of
+  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  * GNU Affero General Public License for more details.
+
+  * You should have received a copy of the GNU Affero General Public License
+  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+  * Copyright (C) 2023-2026 HyperCeiler Contributions
+*/
+package com.sevtinge.hyperceiler.libhook.rules.home.other
 
 import android.net.Uri
 import android.os.Bundle
@@ -22,9 +36,15 @@ import java.util.concurrent.atomic.AtomicBoolean
  * `com.miui.home` native motion chain, so it must not depend on a single rule being selected:
  * both [init] (from `HomePhone`) and [ensureLoadedFromModuleEntry] (from the module entry, before
  * any rule gate) funnel into [ensureLoaded].
+ *
+ * The two `external` functions below are resolved by their JNI symbol names, which encode this
+ * class's fully qualified name. Renaming or moving this object therefore requires updating the
+ * matching `Java_com_sevtinge_hyperceiler_libhook_rules_home_other_NativeHomeHooksOS4_native*`
+ * definitions in `app/src/main/cpp/hyperceiler_home.cpp`. There is no `RegisterNatives` table to
+ * keep the two in sync automatically.
  */
 @Keep
-object NativeHomeHooks {
+object NativeHomeHooksOS4 {
     private external fun nativeConfigure(
         highDeviceLevel: Boolean,
         disablePrestart: Boolean,
